moved all housing market related files into separate package.
