+++ src/main/java/markets/BTLBuyerRecord.java
@@ -1,4 +1,7 @@
-package housing;
+package markets;
+
+import housing.Household;
+import markets.HouseBuyerRecord;
 
 public class BTLBuyerRecord extends HouseBuyerRecord {
 	private static final long serialVersionUID = 5314886568148212605L;
