+++ src/main/java/markets/HousingMarketRecord.java
@@ -1,4 +1,4 @@
-package housing;
+package markets;
 
 import java.io.Serializable;
